Создает структуру модулей
